fix(ReportDownload): Clean up useQueryExportData
